+++ tests/unit/client/reducers/applications.tests.js
