+++ tests/unit/client/reducers/auth.tests.js
