+++ tests/unit/client/reducers/configuration.tests.js
