+++ tests/unit/client/reducers/connections.tests.js
