+++ tests/unit/client/reducers/group.tests.js
