+++ tests/unit/client/reducers/groupMapping.tests.js
