+++ tests/unit/client/reducers/groupMember.tests.js
