+++ tests/unit/client/reducers/groupPicker.tests.js
