+++ tests/unit/client/reducers/groups.tests.js
