+++ tests/unit/client/reducers/permission.tests.js
