+++ tests/unit/client/reducers/permissions.tests.js
