+++ tests/unit/client/reducers/role.tests.js
