+++ tests/unit/client/reducers/ruleStatus.tests.js
