+++ tests/unit/client/reducers/user.tests.js
