+++ tests/unit/client/reducers/userPicker.tests.js
