+++ tests/unit/server/metadata-route.tests.js
