+++ tests/unit/server/roles-route.tests.js
@@ -2,7 +2,7 @@ import Promise from 'bluebird';
 import { expect } from 'chai';
 import { getServerData } from '../server';
 import { getToken } from '../mocks/tokens';
-import config from '../../server/lib/config';
+import config from '../../../server/lib/config';
 
 describe('roles-route', () => {
   const { db, server } = getServerData();
